+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/infrastructure/ProductGridRouter.ts
@@ -1,5 +1,6 @@
 const Router = require('pim/router');
 const DatagridState = require('pim/datagrid/state');
+const userContext = require('pim/user-context');
 
 const PRODUCT_GRID_QUALITY_SCORE_COLUMN = 'data_quality_insights_score';
 
@@ -41,16 +42,14 @@ const buildFilters = (
   if (familyCode) {
     filters = filters.concat([`f[family][value][]=${familyCode}`, 'f[family][type]=in']);
   }
-  if (categoryId) {
-    filters = filters.concat([
-      `f[category][value][treeId]=${rootCategoryId}`,
-      `f[category][value][categoryId]=${categoryId}`,
-      'f[category][type]=1',
-    ]);
-  }
   if (keyIndicator) {
     filters = filters.concat([`f[${keyIndicator}][value]=0`]);
   }
+  filters = filters.concat([
+    `f[category][value][treeId]=${rootCategoryId}`,
+    `f[category][value][categoryId]=${categoryId}`,
+    'f[category][type]=1',
+  ]);
 
   return filters.join('&');
 };
@@ -65,8 +64,8 @@ const redirectToFilteredProductGrid = (
   DatagridState.set('product-grid', {
     columns: productGridColumns.join(','),
     filters: gridFilters,
-    view: '0',
-    initialViewState: '',
+    view: userContext.get('default_product_grid_view') ? userContext.get('default_product_grid_view') : 0,
+    initialViewState: gridFilters,
     scope: channelCode,
   });