+++ src/Akeneo/Platform/Component/Webhook/EventDataBuilderInterface.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Akeneo\Platform\Component\Webhook;
 
 use Akeneo\Platform\Component\EventQueue\BulkEventInterface;
-use Akeneo\Platform\Component\EventQueue\EventInterface;
 use Akeneo\UserManagement\Component\Model\UserInterface;
 
 /**
@@ -14,15 +13,10 @@ use Akeneo\UserManagement\Component\Model\UserInterface;
  */
 interface EventDataBuilderInterface
 {
-    /**
-     * @param EventInterface|BulkEventInterface $event
-     */
-    public function supports(object $event): bool;
+    public function supports(BulkEventInterface $event): bool;
 
     /**
-     * @param EventInterface|BulkEventInterface $event
-     *
      * @return array<mixed> Normalized data.
      */
-    public function build(object $event, UserInterface $user): EventDataCollection;
+    public function build(BulkEventInterface $event, UserInterface $user): EventDataCollection;
 }