+++ tests/back/Pim/Enrichment/Specification/Component/Product/Webhook/ProductModelCreatedAndUpdatedEventDataBuilderSpec.php
@@ -82,13 +82,6 @@ class ProductModelCreatedAndUpdatedEventDataBuilderSpec extends ObjectBehavior
         $this->supports($bulkEvent)->shouldReturn(false);
     }
 
-    public function it_does_not_support_an_individual_event(): void
-    {
-        $event = new ProductModelUpdated(Author::fromNameAndType('julia', Author::TYPE_UI), ['code' => '1']);
-
-        $this->supports($event)->shouldReturn(false);
-    }
-
     public function it_builds_a_bulk_event_of_product_created_and_updated_event(
         ProductQueryBuilderFactoryInterface $pqbFactory,
         GetConnectorProductModels $getConnectorProductModelsQuery,